Disable add layer button if layers limit is reached
