+++ lib/assets/core/javascripts/cartodb3/editor/editor-pane.js
@@ -16,7 +16,6 @@ var checkAndBuildOpts = require('../helpers/required-opts');
 var Infobox = require('../components/infobox/infobox-factory');
 var InfoboxModel = require('../components/infobox/infobox-model');
 var InfoboxCollection = require('../components/infobox/infobox-collection');
-var TipsyTooltipView = require('../components/tipsy-tooltip-view');
 var AppNotifications = require('../app-notifications');
 var zoomToData = require('./map-operations/zoom-to-data');
 
@@ -199,20 +198,9 @@ module.exports = CoreView.extend({
     };
 
     this._mapTabPaneView = createTextLabelsTabPane(tabPaneTabs, tabPaneOptions);
-    this._mapTabPaneView.collection.bind('change:selected', this._updateAddButtonState, this);
 
     this.$el.append(this._mapTabPaneView.render().$el);
     this.addView(this._mapTabPaneView);
-
-    this._updateAddButtonState();
-
-    var tooltip = new TipsyTooltipView({
-      el: this.$('.js-add'),
-      title: function () {
-        return this._tooltipTitle;
-      }.bind(this)
-    });
-    this.addView(tooltip);
 
     this._infoboxState();
 
@@ -226,9 +214,7 @@ module.exports = CoreView.extend({
   },
 
   _initBinds: function () {
-    this.listenTo(this._widgetDefinitionsCollection, 'reset remove add', this._updateAddButtonState);
     this.listenTo(this._editorModel, 'change:edition', this._changeStyle);
-    this.listenTo(this._layerDefinitionsCollection, 'reset remove add', this._updateAddButtonState);
     this.listenTo(this._layerDefinitionsCollection, 'add', this._onLayerAdd);
     this.listenTo(this._layerDefinitionsCollection, 'add remove', this._onLayerCountChange);
     this.listenTo(AppNotifications.getCollection(), 'add', this._infoboxState);
@@ -346,10 +332,8 @@ module.exports = CoreView.extend({
     layersModel.set('label', this._getTranslatedLayersLabel(count, max));
 
     if (count === max) {
-      this._disableAddButton();
       this._infoboxState();
     } else {
-      this._enableAddButton();
       this._infoboxModel.set('state', null);
     }
   },
@@ -395,54 +379,6 @@ module.exports = CoreView.extend({
       count: count,
       maxCount: max
     });
-  },
-
-  _enableAddButton: function () {
-    this.$('.js-add').removeClass('is-disabled');
-  },
-
-  _disableAddButton: function () {
-    this.$('.js-add').addClass('is-disabled');
-  },
-
-  _hideAddButton: function () {
-    this.$('.js-add').addClass('is-hidden');
-  },
-
-  _showAddButton: function () {
-    this.$('.js-add').removeClass('is-hidden');
-  },
-
-  _updateAddButtonState: function () {
-    this._hideAddButton();
-
-    switch (this._mapTabPaneView.getSelectedTabPaneName()) {
-      case 'widgets':
-        this._tooltipTitle = '';
-        if (this._widgetDefinitionsCollection.size()) {
-          this._showAddButton();
-        }
-        this._enableAddButton();
-        break;
-      case LAYERS_TAB_NAME:
-        var count = this._getDataLayerCount();
-        var max = this._getMaxCount();
-
-        if (this._layerDefinitionsCollection.size()) {
-          this._showAddButton();
-        }
-        if (count === max) {
-          this._disableAddButton();
-          this._tooltipTitle = this._getMaxLayerTitle();
-        } else {
-          this._enableAddButton();
-          this._tooltipTitle = '';
-        }
-        break;
-      case 'elements':
-        // TODO: trigger element creation
-        break;
-    }
   },
 
   _share: function () {
