+++ lib/assets/core/javascripts/cartodb3/editor/layers/layers-view.js
@@ -59,6 +59,7 @@ module.exports = CoreView.extend({
 
   _initBinds: function () {
     this.listenTo(this._layerDefinitionsCollection, 'add remove change:id', this.render);
+    this.listenTo(this._layerDefinitionsCollection, 'reset', this._updateAddButtonState);
     this.listenTo(this._layerDefinitionsCollection, 'add', this._createNotification);
   },
 
@@ -68,6 +69,7 @@ module.exports = CoreView.extend({
 
     _.each(this._layerDefinitionsCollection.toArray().reverse(), this._addLayerView, this);
     this._initSortable();
+    this._updateAddButtonState();
     this._initViews();
 
     return this;
@@ -78,8 +80,8 @@ module.exports = CoreView.extend({
       el: this.$('.js-add'),
       gravity: 'w',
       title: function () {
-        return _t('editor.layers.add-layer.tooltip');
-      },
+        return this._tooltipTitle;
+      }.bind(this),
       offset: 8
     });
     this.addView(tooltip);
@@ -142,6 +144,35 @@ module.exports = CoreView.extend({
       beforeStop: this._dontAllowToMoveItemAfterBasemap.bind(this),
       update: this._onSortableUpdate.bind(this)
     });
+  },
+
+  _getDataLayerCount: function () {
+    return this._layerDefinitionsCollection.getNumberOfDataLayers();
+  },
+
+  _getMaxCount: function () {
+    return this._userModel.get('limits').max_layers;
+  },
+
+  _updateAddButtonState: function () {
+    var count = this._getDataLayerCount();
+    var max = this._getMaxCount();
+
+    if (count === max) {
+      this._disableAddButton();
+      this._tooltipTitle = _t('editor.layers.max-layers-infowindow.title');
+    } else {
+      this._enableAddButton();
+      this._tooltipTitle = _t('editor.layers.add-layer.tooltip');
+    }
+  },
+
+  _enableAddButton: function () {
+    this.$('.js-add').removeClass('is-disabled');
+  },
+
+  _disableAddButton: function () {
+    this.$('.js-add').addClass('is-disabled');
   },
 
   _preventPlaceholderAfterBasemap: function (evt, ui) {
