+++ lib/assets/core/test/spec/cartodb3/editor/editor-pane.spec.js
@@ -221,7 +221,7 @@ describe('editor/editor-pane', function () {
   });
 
   it('should have two subviews', function () {
-    expect(_.size(this.view._subviews)).toBe(3);
+    expect(_.size(this.view._subviews)).toBe(2); // [Header, TabPane]
   });
 
   describe('._onLayerCollectionAdd', function () {
@@ -272,25 +272,6 @@ describe('editor/editor-pane', function () {
 
     it('should have no infobox state by default', function () {
       expect(this.view._infoboxModel.get('state')).toBe(null);
-    });
-
-    it('should enable add button by default', function () {
-      expect(this.view.$('.js-add').hasClass('is-disabled')).toBeFalsy();
-    });
-
-    it('should disable the add button when max layers are reached', function () {
-      this.view._layerDefinitionsCollection.add(otherLayer);
-      jasmine.clock().tick(10);
-
-      expect(this.view.$('.js-add').hasClass('is-disabled')).toBeTruthy();
-    });
-
-    it('should re-enable the add button when we remove a layer after max layers are reached', function () {
-      this.view._layerDefinitionsCollection.add(otherLayer);
-      jasmine.clock().tick(10);
-
-      this.view._layerDefinitionsCollection.remove(otherLayer);
-      expect(this.view.$('.js-add').hasClass('is-disabled')).toBeFalsy();
     });
 
     it('should show the max-layers infobox when max layer are reached', function () {
