+++ lib/assets/core/test/spec/cartodb3/editor/layers/layers-view.spec.js
@@ -19,7 +19,12 @@ describe('editor/layers/layers-view', function () {
     var configModel = new ConfigModel({
       base_url: '/u/pepe'
     });
-    var userModel = new UserModel({}, {
+
+    var userModel = new UserModel({
+      limits: {
+        max_layers: 2
+      }
+    }, {
       configModel: configModel
     });
 
@@ -81,8 +86,8 @@ describe('editor/layers/layers-view', function () {
 
     this.view = new LayersView({
       configModel: configModel,
-      userModel: {},
-      pollingModel: {},
+      userModel: userModel,
+      pollingModel: new Backbone.Model(),
       editorModel: new EditorModel(),
       userActions: this.userActions,
       analysis: this.analysis,
@@ -114,6 +119,33 @@ describe('editor/layers/layers-view', function () {
     it('should add all layer views but labels layer', function () {
       var n = this.layerDefinitionsCollection.size();
       expect(this.view._layerViewFactory.createLayerView).toHaveBeenCalledTimes(n - 1);
+    });
+  });
+
+  describe('when max layers limit is reached', function () {
+    it('should disable the add button when max layers are reached', function () {
+      this.layerDefinitionsCollection.add({
+        id: 'l-2',
+        options: {
+          type: 'torque',
+          table_name: 'fee',
+          source: 'a1'
+        }
+      });
+      expect(this.view.$('.js-add').hasClass('is-disabled')).toBeTruthy();
+    });
+
+    it('should re-enable the add button when we remove a layer after max layers are reached', function () {
+      var otherLayer = this.layerDefinitionsCollection.add({
+        id: 'l-2',
+        options: {
+          type: 'torque',
+          table_name: 'fee',
+          source: 'a1'
+        }
+      });
+      this.layerDefinitionsCollection.remove(otherLayer);
+      expect(this.view.$('.js-add').hasClass('is-disabled')).toBeFalsy();
     });
   });
